Adapt the async client to the new API
To match redis-rb API
