+++ lib/3scale/backend/storage_async.rb
@@ -1,3 +1,3 @@
+require '3scale/backend/storage_async/methods'
 require '3scale/backend/storage_async/client'
 require '3scale/backend/storage_async/pipeline'
-require '3scale/backend/storage_async/async_redis'
