+++ lib/3scale/backend/storage_async/pipeline.rb
@@ -5,6 +5,7 @@ module ThreeScale
       # This class accumulates commands and sends several of them in a single
       # request, instead of sending them one by one.
       class Pipeline
+        include Methods
 
         Error = Class.new StandardError
 
@@ -55,11 +56,17 @@ module ThreeScale
           1
         end
 
+        def pipelined(&block)
+          # When running a nested pipeline, we just need to continue
+          # accumulating commands.
+          block.call self
+        end
+
         # Send to redis all the accumulated commands.
         # Returns an array with the result for each command in the same order
         # that they added with .call().
         def run(redis_async_client)
-          responses = redis_async_client.call_pipeline(@commands)
+          responses = collect_responses(redis_async_client)
 
           responses.zip(@commands).map do |resp, cmd|
             command_name = cmd.first.to_s.upcase
@@ -72,6 +79,37 @@ module ThreeScale
               resp
             end
           end
+        end
+
+        private
+
+        def collect_responses(redis_async_client)
+          async_pipe = redis_async_client.pipeline
+          @commands.each do |command|
+            async_pipe.write_request(*command)
+          end
+
+          # Redis returns an answer for each of the commands sent in the
+          # pipeline. But in order to keep compatibility with redis-rb, here, if
+          # there is an error in any of the commands of the pipeline we will
+          # raise an error (the first one that occurred).
+
+          first_err = nil
+
+          res = @commands.size.times.map do
+            begin
+              async_pipe.read_response
+            rescue ::Protocol::Redis::ServerError => e
+              first_err ||= e
+              nil
+            end
+          end
+
+          raise first_err if first_err
+
+          res
+        ensure
+          async_pipe.close
         end
       end
     end
